refactor: modularize runner codebase (Issue #6)
